drop LowLevelLifoSemaphoreBase
