+++ src/mono/System.Private.CoreLib/src/System/Threading/LowLevelLifoSemaphore.Unix.Mono.cs
@@ -5,7 +5,7 @@ using System.Runtime.CompilerServices;
 
 namespace System.Threading
 {
-    internal sealed unsafe partial class LowLevelLifoSemaphore : LowLevelLifoSemaphoreBase, IDisposable
+    internal sealed unsafe partial class LowLevelLifoSemaphore : IDisposable
     {
         private IntPtr lifo_semaphore;
 
@@ -39,7 +39,7 @@ namespace System.Threading
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
         private static extern void ReleaseInternal(IntPtr semaphore, int count);
 
-        protected override void ReleaseCore(int count)
+        private void ReleaseCore(int count)
         {
             ReleaseInternal(lifo_semaphore, count);
         }
